+++ ChangeLog.md
@@ -1,5 +1,7 @@
 ### Summary
 
-- Fixed the offset for `m_iIDEntIndex`
+- Added information about the [Help Center](https://violetwing.featurebase.app/en/help).
+- Completely rewrites the FAQ tab.
+- Displaying information about the latest CS2 patch.
 
-![Downloads](https://img.shields.io/github/downloads/Jesewe/VioletWing/v1.2.8.2/total?style=for-the-badge&logo=github&color=D5006D) ![Platforms](https://img.shields.io/badge/platform-Windows-blue?style=for-the-badge&color=D5006D) ![License](https://img.shields.io/github/license/jesewe/cs2-triggerbot?style=for-the-badge&color=D5006D)
+![Downloads](https://img.shields.io/github/downloads/Jesewe/VioletWing/v1.2.8.3/total?style=for-the-badge&logo=github&color=D5006D) ![Platforms](https://img.shields.io/badge/platform-Windows-blue?style=for-the-badge&color=D5006D) ![License](https://img.shields.io/github/license/jesewe/cs2-triggerbot?style=for-the-badge&color=D5006D)